+++ module/actor/entity.js
@@ -834,6 +834,7 @@ export class ActorPF extends Actor {
     attackData["data.duration.units"] = "inst";
     attackData["data.range.units"] = "melee";
     attackData["data.broken"] = item.data.data.broken;
+    attackData["data.range.maxIncrements"] = item.data.data.weaponData.maxRangeIncrements;
     attackData["img"] = item.data.img;
 
     // Add additional attacks